Make AdditionalTestOutputPlugin work without on device dir
If on-device directory is not specified, the plugin only
copies files from the TestStorage service.

Bug: 195582093
Test: unit tests and tested manually

+++ utp/android-test-plugin-host-additional-test-output/src/main/java/com/android/tools/utp/plugins/host/additionaltestoutput/AndroidAdditionalTestOutputPlugin.kt
@@ -100,8 +100,10 @@ class AndroidAdditionalTestOutputPlugin(private val logger: Logger = getLogger()
      */
     private fun createEmptyDirectoryOnDevice(deviceController: DeviceController) {
         val dir = config.additionalOutputDirectoryOnDevice
-        deviceController.deviceShellAndCheckSuccess("rm -rf \"${dir}\"")
-        deviceController.deviceShellAndCheckSuccess("mkdir -p \"${dir}\"")
+        if (dir.isNotBlank()) {
+            deviceController.deviceShellAndCheckSuccess("rm -rf \"${dir}\"")
+            deviceController.deviceShellAndCheckSuccess("mkdir -p \"${dir}\"")
+        }
 
         if (deviceController.isTestServiceInstalled()) {
             deviceController.deviceShellAndCheckSuccess(
@@ -247,6 +249,10 @@ class AndroidAdditionalTestOutputPlugin(private val logger: Logger = getLogger()
 
     private fun copyAdditionalTestOutputsFromDeviceToHost(deviceController: DeviceController) {
         val deviceDir = config.additionalOutputDirectoryOnDevice
+        if (deviceDir.isBlank()) {
+            return
+        }
+
         val hostDir = File(config.additionalOutputDirectoryOnHost).absolutePath
         copyFilesFromDeviceToHost(deviceController, deviceDir, hostDir)
     }
